Add files via upload
Edited the intervention grid function to input raw data and output the standardized intervention values, edited the intervention code to accept and return standardized values, and then added the inverse functions to inverse the standardized value back to the raw values for plotting. Also added the code to load all of the SCMs from the causal modeling step.